+++ lib/carrierwave-aliyun.rb
@@ -1,9 +1,9 @@
 require 'carrierwave/storage/aliyun'
+require 'carrierwave/storage/aliyun_file'
+require 'carrierwave/aliyun/bucket'
+require 'carrierwave/aliyun/version'
 require 'carrierwave/aliyun/configuration'
 
-CarrierWave.configure do |config|
-  config.storage_engines.merge!({ aliyun: 'CarrierWave::Storage::Aliyun' })
-end
 CarrierWave::Uploader::Base.send(:include, CarrierWave::Aliyun::Configuration)
 
 if CarrierWave::VERSION <= '0.11.0'